requirements.txt
gym-retro==0.8.0
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.0.74
neat-python==0.92 